moving useFormatBalance to the api module
